+++ packages/ragbits-evaluate/src/ragbits/evaluate/optimizer.py
@@ -9,8 +9,9 @@ from pydantic import BaseModel
 
 from ragbits.core.utils.config_handling import WithConstructionConfig, import_by_path
 from ragbits.evaluate.dataloaders.base import DataLoader
-from ragbits.evaluate.evaluator import Evaluator, EvaluatorConfig
+from ragbits.evaluate.evaluator import Evaluator
 from ragbits.evaluate.metrics.base import MetricSet
+from ragbits.evaluate.evaluator import EvaluatorConfig
 from ragbits.evaluate.pipelines.base import EvaluationPipeline
 from ragbits.evaluate.utils import setup_optuna_neptune_callback
 
@@ -74,6 +75,7 @@ class Optimizer(WithConstructionConfig):
             metrics=metrics,
             dataloader=dataloader,
             callbacks=callbacks,
+            schema_config=evaluator_config.schema_config,
         )
 
     def optimize(
@@ -83,6 +85,7 @@ class Optimizer(WithConstructionConfig):
         dataloader: DataLoader,
         metrics: MetricSet,
         callbacks: list[Callable] | None = None,
+        schema_config: dict | None = None,
     ) -> list[tuple[dict, float, dict[str, float]]]:
         """
         Runs the optimization process for given parameters.
@@ -105,6 +108,7 @@ class Optimizer(WithConstructionConfig):
                 pipeline_config=pipeline_config,
                 dataloader=dataloader,
                 metrics=metrics,
+                schema_config=schema_config,
             )
 
         study = optuna.create_study(direction=self.direction)
@@ -132,16 +136,17 @@ class Optimizer(WithConstructionConfig):
         pipeline_config: dict,
         dataloader: DataLoader,
         metrics: MetricSet,
+        schema_config: dict | None = None,
     ) -> float:
         """
         Runs a single experiment.
         """
-        evaluator = Evaluator()
         event_loop = asyncio.get_event_loop()
 
         score = 1e16 if self.direction == "maximize" else -1e16
         metrics_values = None
         config_for_trial = None
+        evaluator: Evaluator = Evaluator(schema_config=schema_config, pipeline_type=pipeline_class.configuration_key)
 
         for attempt in range(1, self.max_retries_for_trial + 1):
             try: